+++ packages/components/progress-stepper/examples/ProgressStepperHorizontalLabelExample.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { ProgressStepper } from '@contentful/f36-progress-stepper';
+
+export default function ProgressStepperHorizontalLabelExample() {
+  return (
+    <ProgressStepper
+      activeStep={2}
+      ariaLabel="Horizontal progress stepper with labels"
+    >
+      <ProgressStepper.Step state="complete" labelText="Label 1" />
+      <ProgressStepper.Step state="complete" labelText="Label 2" />
+      <ProgressStepper.Step state="active" labelText="Label 3" />
+      <ProgressStepper.Step labelText="Label 4" />
+      <ProgressStepper.Step labelText="Label 5" />
+    </ProgressStepper>
+  );
+}