+++ packages/forma-36-codemod/utils/updateComponentsToImport.js
@@ -1,3 +1,9 @@
+/**
+ * Update the list of components to import in the source code
+ * @param {*} componentsToImport - array of components to import
+ * @param {*} newImports - array or string of new imports added to source
+ * @returns Updated list of compoentns to import
+ */
 module.exports.updateComponentsToImport = (componentsToImport, newImports) => {
   let _imports = newImports;
   let allComponents = componentsToImport;