+++ packages/components/avatar/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 4.67.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.67.2
+  - @contentful/f36-menu@4.67.2
+  - @contentful/f36-tooltip@4.67.2
+  - @contentful/f36-image@4.67.2
+
 ## 4.67.1
 
 ### Patch Changes