+++ packages/components/navbar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @contentful/f36-navbar
 
+## 4.67.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @contentful/f36-core@4.67.2
+  - @contentful/f36-icon@4.67.2
+  - @contentful/f36-menu@4.67.2
+  - @contentful/f36-skeleton@4.67.2
+  - @contentful/f36-avatar@4.67.2
+
 ## 4.67.1
 
 ### Patch Changes